User commit

+++ src/main/java/com/example/demo/entity/ValidationEntity.java
@@ -27,7 +27,7 @@ public Integer getId(){
     return id;
 }
 
-public void setId(Integer id){
+public void setId(Long id){
     this.id=id;
 }
 public String getUsername(){
